# tequilapi/docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The purpose of this documentation is to provide developers an insight of how to\ninteract with Mysterium Node via Tequila API.\nThis should demonstrate all the possible API calls with described parameters and responses.",
    "title": "Tequila API",
    "version": "dev"
  },
  "host": "127.0.0.1:4050",
  "paths": {
...
    "/connection": {
      "get": {
        "description": "Returns status of current connection",
        "tags": [
          "Connection"
        ],
        "summary": "Returns connection status",
        "operationId": "connectionStatus",
        "responses": {
          "200": {
            "description": "Status",
            "schema": {
              "$ref": "#/definitions/ConnectionInfoDTO"
            }
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      },
      "put": {
        "description": "Consumer opens connection to provider",
        "tags": [
          "Connection"
        ],
        "summary": "Starts new connection",
        "operationId": "connectionCreate",
        "parameters": [
          {
            "description": "Parameters in body (consumer_id, provider_id, service_type) required for creating new connection",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ConnectionCreateRequestDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Connection started",
            "schema": {
              "$ref": "#/definitions/ConnectionInfoDTO"
            }
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "422": {
            "description": "Unable to process the request at this point",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      },
      "delete": {
        "description": "Stops current connection",
        "tags": [
          "Connection"
        ],
        "summary": "Stops connection",
        "operationId": "connectionCancel",
        "responses": {
          "202": {
            "description": "Connection stopped"
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "422": {
            "description": "Unable to process the request at this point (e.g. no active connection exists)",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
...
    "/connection/traffic": {
      "get": {
        "description": "Returns traffic information about requested connection",
        "tags": [
          "Connection"
        ],
        "summary": "Returns connection traffic information",
        "operationId": "connectionTraffic",
        "responses": {
          "200": {
            "description": "Connection traffic",
            "schema": {
              "$ref": "#/definitions/ConnectionTrafficDTO"
            }
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
...
    "/exchange/myst/{currency}": {
      "get": {
        "description": "Returns the myst price in the given currency (dai is deprecated)",
        "tags": [
          "Exchange"
        ],
        "summary": "Returns the myst price in the given currency",
        "operationId": "ExchangeMyst",
        "parameters": [
          {
            "type": "string",
            "description": "Currency to which MYST is converted",
            "name": "currency",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "MYST price in given currency",
            "schema": {
              "$ref": "#/definitions/CurrencyExchangeDTO"
            }
          },
          "404": {
            "description": "Currency is not supported",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "APIError": {
      "type": "object",
      "title": "APIError represents an error response from REST API service.",
      "properties": {
        "error": {
          "$ref": "#/definitions/Err"
        },
        "path": {
          "type": "string",
          "x-go-name": "Path"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-name": "apiErrorSwagger",
      "x-go-package": "github.com/mysteriumnetwork/go-rest/apierror"
    },
...
    "AccessPolicy": {
      "description": "AccessPolicy represents the access controls for proposal",
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "source": {
          "type": "string",
          "x-go-name": "Source"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/market"
    },
...
    "ConnectOptionsDTO": {
      "description": "ConnectOptions holds tequilapi connect options",
      "type": "object",
      "properties": {
        "dns": {
          "$ref": "#/definitions/DNSOption"
        },
        "kill_switch": {
          "description": "kill switch option restricting communication only through VPN",
          "type": "boolean",
          "x-go-name": "DisableKillSwitch",
          "example": true
        },
        "proxy_port": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ProxyPort"
        }
      },
      "x-go-name": "ConnectOptions",
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
    "ConnectionCreateFilter": {
      "description": "ConnectionCreateFilter describes filter for the connection request to lookup\nfor a requested proposals based on specified params.",
      "type": "object",
      "properties": {
        "country_code": {
          "type": "string",
          "x-go-name": "CountryCode"
        },
        "include_monitoring_failed": {
          "type": "boolean",
          "x-go-name": "IncludeMonitoringFailed"
        },
        "ip_type": {
          "type": "string",
          "x-go-name": "IPType"
        },
        "providers": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Providers"
        },
        "sort_by": {
          "type": "string",
          "x-go-name": "SortBy"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
    "ConnectionCreateRequestDTO": {
      "type": "object",
      "title": "ConnectionCreateRequest request used to start a connection.",
      "required": [
        "consumer_id",
        "provider_id"
      ],
      "properties": {
        "connect_options": {
          "$ref": "#/definitions/ConnectOptionsDTO"
        },
        "consumer_id": {
          "description": "consumer identity",
          "type": "string",
          "x-go-name": "ConsumerID",
          "example": "0x0000000000000000000000000000000000000001"
        },
        "filter": {
          "$ref": "#/definitions/ConnectionCreateFilter"
        },
        "hermes_id": {
          "description": "hermes identity",
          "type": "string",
          "x-go-name": "HermesID",
          "example": "0x0000000000000000000000000000000000000003"
        },
        "provider_id": {
          "description": "provider identity",
          "type": "string",
          "x-go-name": "ProviderID",
          "example": "0x0000000000000000000000000000000000000002"
        },
        "service_type": {
          "description": "service type. Possible values are \"openvpn\", \"wireguard\" and \"noop\"",
          "type": "string",
          "default": "openvpn",
          "x-go-name": "ServiceType",
          "example": "openvpn"
        }
      },
      "x-go-name": "ConnectionCreateRequest",
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "ConnectionInfoDTO": {
      "type": "object",
      "title": "ConnectionInfoDTO holds partial consumer connection details.",
      "properties": {
        "consumer_id": {
          "type": "string",
          "x-go-name": "ConsumerID",
          "example": "0x00"
        },
        "hermes_id": {
          "type": "string",
          "x-go-name": "HermesID",
          "example": "0x00"
        },
        "proposal": {
          "$ref": "#/definitions/ProposalDTO"
        },
        "session_id": {
          "type": "string",
          "x-go-name": "SessionID",
          "example": "4cfb0324-daf6-4ad8-448b-e61fe0a1f918"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status",
          "example": "Connected"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "ConnectionTrafficDTO": {
      "type": "object",
      "title": "ConnectionTrafficDTO holds consumer connection traffic information.",
      "properties": {
        "bytes_received": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "BytesReceived",
          "example": 1024
        },
        "bytes_sent": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "BytesSent",
          "example": 1024
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "CurrencyExchangeDTO": {
      "type": "object",
      "title": "CurrencyExchangeDTO the value of a given currency.",
      "properties": {
        "amount": {
          "type": "number",
          "format": "double",
          "x-go-name": "Amount"
        },
        "currency": {
          "type": "string",
          "x-go-name": "Currency"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "DNSOption": {
      "description": "DNSOption defines DNS server selection strategy for consumer",
      "type": "string",
      "x-go-package": "github.com/mysteriumnetwork/node/core/connection"
    },
...
    "Err": {
      "type": "object",
      "title": "Err contains the error details.",
      "properties": {
        "code": {
          "type": "string",
          "x-go-name": "Code"
        },
        "detail": {
          "type": "string",
          "x-go-name": "Detail"
        },
        "fields": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/FieldError"
          },
          "x-go-name": "Fields"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/go-rest/apierror"
    },
...
    "FieldError": {
      "type": "object",
      "title": "FieldError contains the reason why a field failed validation.",
      "properties": {
        "code": {
          "type": "string",
          "x-go-name": "Code"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/go-rest/apierror"
    },
...
    "Price": {
      "type": "object",
      "title": "Price represents the service price.",
      "properties": {
        "currency": {
          "type": "string",
          "x-go-name": "Currency"
        },
        "per_gib": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "PerGiB"
        },
        "per_gib_tokens": {
          "$ref": "#/definitions/Tokens"
        },
        "per_hour": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "PerHour"
        },
        "per_hour_tokens": {
          "$ref": "#/definitions/Tokens"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
    "ProposalDTO": {
      "type": "object",
      "title": "ProposalDTO holds service proposal details.",
      "properties": {
        "access_policies": {
          "description": "AccessPolicies",
          "type": "array",
          "items": {
            "$ref": "#/definitions/AccessPolicy"
          },
          "x-go-name": "AccessPolicies"
        },
        "compatibility": {
          "description": "Compatibility level.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Compatibility"
        },
        "format": {
          "description": "Proposal format.",
          "type": "string",
          "x-go-name": "Format"
        },
        "location": {
          "$ref": "#/definitions/ServiceLocationDTO"
        },
        "price": {
          "$ref": "#/definitions/Price"
        },
        "provider_id": {
          "description": "provider who offers service",
          "type": "string",
          "x-go-name": "ProviderID",
          "example": "0x0000000000000000000000000000000000000001"
        },
        "quality": {
          "$ref": "#/definitions/Quality"
        },
        "service_type": {
          "description": "type of service provider offers",
          "type": "string",
          "x-go-name": "ServiceType",
          "example": "openvpn"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "Quality": {
      "type": "object",
      "title": "Quality holds proposal quality metrics.",
      "properties": {
        "bandwidth": {
          "type": "number",
          "format": "double",
          "x-go-name": "Bandwidth"
        },
        "latency": {
          "type": "number",
          "format": "double",
          "x-go-name": "Latency"
        },
        "quality": {
          "type": "number",
          "format": "double",
          "x-go-name": "Quality"
        },
        "uptime": {
          "type": "number",
          "format": "double",
          "x-go-name": "Uptime"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "ServiceLocationDTO": {
      "type": "object",
      "title": "ServiceLocationDTO holds service location metadata.",
      "properties": {
        "asn": {
          "description": "Autonomous System Number",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ASN",
          "example": 1
        },
        "city": {
          "type": "string",
          "x-go-name": "City",
          "example": "Amsterdam"
        },
        "continent": {
          "type": "string",
          "x-go-name": "Continent",
          "example": "EU"
        },
        "country": {
          "type": "string",
          "x-go-name": "Country",
          "example": "NL"
        },
        "ip_type": {
          "type": "string",
          "x-go-name": "IPType",
          "example": "residential"
        },
        "isp": {
          "type": "string",
          "x-go-name": "ISP",
          "example": "Telia Lietuva, AB"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "Tokens": {
      "description": "Tokens a common response for ethereum blockchain monetary amount",
      "type": "object",
      "properties": {
        "ether": {
          "type": "string",
          "x-go-name": "Ether"
        },
        "human": {
          "type": "string",
          "x-go-name": "Human"
        },
        "wei": {
          "type": "string",
          "x-go-name": "Wei"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
  }
}